backend/requirements.txt
fastapi>=0.105.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2
sqlalchemy>=2.0.23
pdfplumber>=0.10.2
python-docx>=0.8.11
beautifulsoup4>=4.12.2
pytest>=7.4.3
httpx>=0.25.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
python-dotenv>=1.0.0
lxml>=4.9.3
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.1.0
# torch>=2.1.1+cu118  # For CUDA 11.8 (RTX 4050 compatible)
sentence-transformers>=2.2.2
langchain-community>=0.0.10
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tiktoken>=0.5.2
pandas>=2.1.4
scipy>=1.11.4
tqdm>=4.66.1
rich>=13.7.0
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
